standardized use of the AccountNumber component across the tx drilldown, drawer, and secure send modal
