+++ packages/mobile/src/components/AccountNumberCard.tsx
@@ -1,42 +1,23 @@
 import Card from '@celo/react-components/components/Card'
-import colors from '@celo/react-components/styles/colors.v2'
-import fontStyles from '@celo/react-components/styles/fonts.v2'
-import { getAddressChunks } from '@celo/utils/src/address'
 import React from 'react'
-import { StyleSheet, Text } from 'react-native'
+import { StyleSheet, ViewStyle } from 'react-native'
+import AccountNumber from 'src/components/AccountNumber'
 
 interface Props {
   address: string
+  style?: ViewStyle
 }
 
-export default function AccountNumberCard({ address }: Props) {
-  // Turns '0xce10ce10ce10ce10ce10ce10ce10ce10ce10ce10'
-  // into 'ce10 ce10 ce10 ce10 ce10 ce10 ce10 ce10 ce10 ce10'
-  const addressChunks = ['0x', ...getAddressChunks(address)]
+export default function AccountNumberCard({ address, style }: Props) {
   return (
-    <Card style={styles.container} rounded={true}>
-      <Text style={[styles.text, styles.topText]}>{addressChunks.slice(0, 6).join(' ')}</Text>
-      <Text style={[styles.text, styles.bottomText]}>{addressChunks.slice(6).join(' ')}</Text>
+    <Card style={[styles.container, style]} rounded={true}>
+      <AccountNumber address={address} />
     </Card>
   )
 }
 
 const styles = StyleSheet.create({
   container: {
-    height: 62,
-    width: 235,
-    paddingHorizontal: 16,
     justifyContent: 'center',
-  },
-  text: {
-    ...fontStyles.small,
-    color: colors.gray4,
-  },
-  topText: {
-    textAlign: 'left',
-  },
-  bottomText: {
-    paddingTop: 4,
-    textAlign: 'right',
   },
 })
